Rename all AD916x support to AD9166
