+++ adi/__init__.py
@@ -33,7 +33,6 @@
 
 from adi.ad717x import ad717x
 from adi.ad719x import ad719x
-from adi.ad916x import ad9166
 from adi.ad936x import Pluto, ad9361, ad9363, ad9364
 from adi.ad4110 import ad4110
 from adi.ad4130 import ad4130
@@ -52,6 +51,7 @@ from adi.ad9094 import ad9094
 from adi.ad9136 import ad9136
 from adi.ad9144 import ad9144
 from adi.ad9152 import ad9152
+from adi.ad9166 import ad9166
 from adi.ad9172 import ad9172
 from adi.ad9250 import ad9250
 from adi.ad9265 import ad9265
