+++ doc/source/devices/adi.ad9166.rst
@@ -1,7 +1,7 @@
 adi.ad9166 module
 =================
 
-.. automodule:: adi.ad916x
+.. automodule:: adi.ad9166
    :members:
    :undoc-members:
    :show-inheritance:
